requirement.txt
easyocr==1.7.2
opencv-python==4.12.0.88
pillow==12.0.0
numpy==2.2.6
deep-translator==1.11.4
PyQt5==5.15.11
pyqt5-qt5==5.15.2
pyqt5-sip==12.17.1
langdetect==1.0.9
fuzzywuzzy==0.18.0
python-Levenshtein==0.27.3
torch==2.9.1
torchvision==0.24.1